chore: move spawn-process to utils directory
Reason: it's shared between copy-modles and generate-clients scripts
