+++ scripts/generate-clients/code-prettify.js
@@ -1,9 +1,12 @@
 // @ts-check
-const { spawnProcess } = require("./spawn-process");
+const { spawnProcess } = require("../utils/spawn-process");
 const path = require("path");
 
 const prettifyCode = async (dir) => {
-  await spawnProcess(path.join(__dirname, "..", "..", "node_modules", ".bin", "prettier"), ["--write", `${dir}/**/*.{ts,js,md,json}`]);
+  await spawnProcess(path.join(__dirname, "..", "..", "node_modules", ".bin", "prettier"), [
+    "--write",
+    `${dir}/**/*.{ts,js,md,json}`,
+  ]);
 };
 
 module.exports = {
